Add existing project files to Git
