+++ Brechó/Brechó.csproj
@@ -47,24 +47,24 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CadastroDeCliente.cs" />
-    <Compile Include="Cliente.cs" />
-    <Compile Include="FrmFornecedorCadastro.cs">
+    <Compile Include="Classes\CadastroDeCliente.cs" />
+    <Compile Include="Classes\Cliente.cs" />
+    <Compile Include="Classes\FrmFornecedorCadastro.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="FrmFornecedorCadastro.Designer.cs">
+    <Compile Include="Classes\FrmFornecedorCadastro.Designer.cs">
       <DependentUpon>FrmFornecedorCadastro.cs</DependentUpon>
     </Compile>
-    <Compile Include="frmInico.cs">
+    <Compile Include="Formularios\frmInico.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="frmInico.Designer.cs">
+    <Compile Include="Formularios\frmInico.Designer.cs">
       <DependentUpon>frmInico.cs</DependentUpon>
     </Compile>
-    <Compile Include="frmMetododePagamento.cs">
+    <Compile Include="Formularios\frmMetododePagamento.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="frmMetododePagamento.Designer.cs">
+    <Compile Include="Formularios\frmMetododePagamento.Designer.cs">
       <DependentUpon>frmMetododePagamento.cs</DependentUpon>
     </Compile>
     <Compile Include="msdbDataSet.Designer.cs">
@@ -72,16 +72,16 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>msdbDataSet.xsd</DependentUpon>
     </Compile>
-    <Compile Include="NovaVenda.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="Classes\NovaVenda.cs" />
+    <Compile Include="Classes\Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="FrmFornecedorCadastro.resx">
+    <EmbeddedResource Include="Classes\FrmFornecedorCadastro.resx">
       <DependentUpon>FrmFornecedorCadastro.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="frmInico.resx">
+    <EmbeddedResource Include="Formularios\frmInico.resx">
       <DependentUpon>frmInico.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="frmMetododePagamento.resx">
+    <EmbeddedResource Include="Formularios\frmMetododePagamento.resx">
       <DependentUpon>frmMetododePagamento.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
